+++ src/udtHelpers.ts
@@ -55,10 +55,7 @@ export function commandToPacketString(command: Uint8Array): string {
   if (command.length === 0) {
     return "[]";
   }
-  let cmdStr = "[";
-  command.forEach(n => cmdStr += n.toString(16) + ",");
-  cmdStr = cmdStr.slice(0, -1) + "]";
-  return cmdStr;
+  return `[${Array.from(command).map(n => n.toString(16).padStart(2, '0')).join(',')}]`;
 }
 
 /**